SDK-763: remove abstract cross chain validators and their body to validate classes; the validation will be performed by the two concrete validators in the SidechainState and by the custom validators that will be defined and injected in the application layer
